requirements.txt
docker
python-dotenv